requirements.txt
fastapi
uvicorn
python-dotenv
yt-dlp
google-generativeai
pydantic
duckduckgo-search